Keylogger with Email Capabilty
This repository contains a Python-based keylogger that captures keystrokes and sends periodic email reports containing the captured data. The keylogger logs user input and stores it in a local text file, which is then processed and sent via email when a certain file size is reached. The email report includes the number of keyword occurrences detected in the captured text, providing insights into specific keywords typed by the user.